[#173] refactor: 로고 컴포넌트 분리
