+++ src/components/Authentication/Join/LogoContainer/Styles.ts
@@ -1,0 +1,25 @@
+import styled from 'styled-components';
+import * as I from '../../../../assets/Icons';
+
+export const LogoContainer = styled.div`
+`;
+
+export const ZimeetLogo = styled(I.ZimeetLogo)`
+  width: 20%;
+  height: 20%;
+`;
+
+export const LogoTitle = styled.p`
+  font-weight: 700;
+  font-family: 'Pretendard';
+  font-size: 17px;
+  color: #000000;
+  padding: 2% 0;
+`;
+
+export const LogoText = styled.p`
+  font-size: 14px;
+  color: #707070;
+  padding: 1% 0;
+  font-family: Pretendard;
+`;
